requirements.txt
streamlit>=1.36
torch>=2.2
torchvision>=0.17
pillow>=10.3
numpy>=1.24
pandas>=2.2
altair>=5.2
matplotlib>=3.8
